+++ OsuDiffCalc/FileProcessor/BeatmapObjects/Hitcircle.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace OsuDiffCalc.FileProcessor.BeatmapObjects {
 	class Hitcircle : HitObject {
-		public Hitcircle(float x, float y, double startTime) : base(x, y, startTime, startTime) {
+		public Hitcircle(float x, float y, int startTime) : base(x, y, startTime, startTime) {
 		}
 	}
 }
